[aotinductor] Turn clang warning ignored-optimization-argument into error (#112008)
Now we compile the generated wrapper C++ code with clang in fbcode.
When the Model's run_impl function is too large, clang will issue
a warning like:

  Function foo is too big to optimize [-Wignored-optimization-argument]

and compile the code without any optimization.

I think we may want to be more proactive in such cases. If the
generated C++ code is too complex or too large to be optimized,
we would like to be notified loudly with errors, so that we
would figure out ways to address the issue.

Later if we feel that turning this warning into an error is too
aggressive, we would add a config to disable it.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/112008
Approved by: https://github.com/desertfire, https://github.com/htyu